+++ src/Revit.Linter.ThemeManaging/Abstractions/Services/IThemeService.cs
@@ -1,5 +1,4 @@
 using System.Windows;
-using System.Windows.Media;
 
 namespace Revit.Linter.ThemeManaging.Abstractions.Services;
 
@@ -7,5 +6,5 @@ public interface IThemeService
 {
     void Register(FrameworkElement element);
 
-    void ChangeTheme(bool isDarkTheme, Color? backgroundColor = null);
+    void ChangeTheme(bool isDarkTheme);
 }